backend/requirements-test.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-bdd>=7.0.0
httpx>=0.27.0
PyYAML>=6.0.0
